refactor: extract price refresh to a new composable
